Remove the buildtime_bindgen feature
I'm not sure how to get it working, and we don't need it.

I've verfied that all the current bindings are totally independant of the
environment: croaring doesn't use any #ifdefs to change any types defined in
the APIs. All types in the API are the same for all architectures (that's not
to say the size of the types are the same, for instance, size_t is used, but
that is translated to the single type usize.

Disable bindgen layout tests. These are archetecture specific (and really are
just testing bindgen's own implementation.